Create Translations with SiteAccess
This would have so helpful to me, it joins the 3 main dots of translation, siteaccess and permissions in one go, linking to the rel. full pages when required. I'm not sure if i have the naming of SiteAccess correct in all cases though :) Not sure if the siteaccess is only the section siteaccess or also the sections under system?