Create Translations with SiteAccess (#52)
This would have so helpful to me, it joins the 3 main dots of translation, siteaccess and permissions in one go, linking to the rel. full pages when required. I'm not sure if i have the naming of SiteAccess correct in all cases though :) Not sure if the siteaccess is only the section siteaccess or also the sections under system?

Update and rename Translations with SiteAccess to How-to create Translations with SiteAccess

Copy changes added

Title change

In accordance with the comments on the commit, the title on 2nd review could have been interpreted incorrectly.

Proof and adapt to style guide