Update requirements for Ubuntu 18.04 & apache 2.4
1. Update requirements for v2.2 to reflect the new Ubuntu 18.04LTS release
2. Clarify that we now recommend event MPM for use with Apache, which implies usage of php-fpm for fastcgi setup
3. Some other minor fixes for v2


#### DOC / QA
Point 2 has an impact on you both, install doc/testing for linux should be updated to reflect usage of Apache event MPM, instead of `prefork` + `mod_php` which is done here: https://github.com/ezsystems/developer-documentation/blob/cdeec99192d66af91d6f105cf90f704236b152f1/docs/getting_started/install_manually.md#8-server-configuration

TL;DR on why; prefork uses more memory as it loads php for all requests, even static once (css, js, ..), and on top of that each child process is only able to handle one request at a time. event MPM solves both of those issues.

A bit more background: For many years prefork was the way to configure PHP, where php lived as a module in apache. Back in 2010 php-fpm came, a official way to run PHP via fastcgi which was the only way to run PHP in nginx and lighttpd, Apache 2.4 actually has become the default in all our supported linux distroes over the last few years, and with it event MPM was made stable and the preferred way to run PHP in production. So we are a bit behind the curve on this one, simply because we have mainly been using nginx for the last few years ourselves.

#### Engineering

Before this can be merged, inline doc and vhost config also need to be updated for this, if someone happens to have a working Apache 2.4 setup with php-fpm, maybe they can share that as a starting point. Ref: https://github.com/ezsystems/ezplatform/tree/master/doc/apache2

So seems like this needs to be added in one of the upcoming sprints before we enter 2.2 certification.