Fixed bad SiteAccess configuration example
Configuration example in SiteAccess Matching section is wrong, as it is missing `admin_group` and assigns `ezdemo_site_admin` SiteAccess to `ezdemo_site_group`. Also, a description of the `ezpublish.siteaccess.groups` parameter is wrong, as it is not optional (because `admin_group` has to be defined) and it is missing information about `admin_group`.