2.5.6 introduce password expiration and not recovery
In 2.5.6 we introduce a new column in the ezsuer DB table which is in defined in dbupdate-7.5.4-to-7.5.5.sql file. This column will be used to save current time when user password is changed. The Password Recovery process is using the same process like forgot password introduced in previous release.